backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.8.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.1

# Security
slowapi==0.1.9
python-decouple==3.8
secure==0.3.0

# AI/ML
openai==1.14.0
anthropic==0.7.0
tiktoken==0.7.0
numpy==1.26.2

# Development
pytest==8.1.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
ruff==0.1.6
httpx==0.27.0
factory-boy==3.3.0
faker==20.1.0
pre-commit==4.2.0

# Additional
redis==5.0.1
aioredis==2.0.1
python-dateutil==2.8.2
reportlab==4.0.7

# Monitoring
sentry-sdk[fastapi]==1.40.0
structlog==24.1.0
